requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
python-docx==1.1.0
pdfplumber==0.10.3
Jinja2==3.1.2